// src/app/api/generated/model/create-bloque-request.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Datos para dar de alta un bloque recurrente de disponibilidad
 */
export interface CreateBloqueRequest { 
    /**
     * Dia de la semana en formato ISO-8601: lunes = 1 .. domingo = 7
     */
    diaSemana: number;
    /**
     * Hora local de inicio, en la zona de la sede. NUNCA un instante UTC: un lunes 09:00 tiene que seguir siendo 09:00 despues de un cambio de huso o de horario de verano
     */
    horaDesde: string;
    /**
     * Hora local de fin, EXCLUSIVA. Un bloque que llega a la medianoche se manda como 24:00, que es el unico valor que expresa el final del dia en un extremo exclusivo: 00:00 significaria el principio. Nunca cruza al dia siguiente
     */
    horaHasta: string;
    /**
     * Primer dia en que el bloque rige. Omitirlo significa HOY en la zona de la sede, que es lo que quiere decir quien carga el horario de un profesional que ya esta atendiendo
     */
    vigenciaDesde?: string;
    /**
     * Primer dia en que el bloque ya NO rige, EXCLUSIVO. Omitirlo significa sin fin previsto, que es el caso normal
     */
    vigenciaHasta?: string;
}